requirements.txt
python-telegram-bot==21.6
requests==2.32.3
google-search-results==2.4.2
